moved trie benchmarks to normal test files
