+++ trie/bin_test.go
@@ -103,3 +103,84 @@ func TestBinBatch(t *testing.T) {
 		t.Fatalf("root hash not zero: %x != %x", hash, zero)
 	}
 }
+
+func BenchmarkBinInsert(b *testing.B) {
+	t := NewBin()
+	keys := make([][]byte, 0, b.N)
+	hashes := make([][]byte, 0, b.N)
+	for i := 0; i < b.N; i++ {
+		key := make([]byte, 32)
+		hash := make([]byte, 32)
+		_, _ = rand.Read(key)
+		_, _ = rand.Read(hash)
+		keys = append(keys, key)
+		hashes = append(hashes, hash)
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		t.MustPut(keys[i], hashes[i])
+	}
+}
+
+func BenchmarkBinRetrieve(b *testing.B) {
+	t := NewBin()
+	keys := make([][]byte, 0, b.N)
+	hashes := make([][]byte, 0, b.N)
+	for i := 0; i < b.N; i++ {
+		key := make([]byte, 32)
+		hash := make([]byte, 32)
+		_, _ = rand.Read(key)
+		_, _ = rand.Read(hash)
+		keys = append(keys, key)
+		hashes = append(hashes, hash)
+	}
+	for i := 0; i < b.N; i++ {
+		t.MustPut(keys[i], hashes[i])
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		t.Get(keys[i])
+	}
+}
+
+func BenchmarkBinDelete(b *testing.B) {
+	t := NewBin()
+	keys := make([][]byte, 0, b.N)
+	hashes := make([][]byte, 0, b.N)
+	for i := 0; i < b.N; i++ {
+		key := make([]byte, 32)
+		hash := make([]byte, 32)
+		_, _ = rand.Read(key)
+		_, _ = rand.Read(hash)
+		keys = append(keys, key)
+		hashes = append(hashes, hash)
+	}
+	for i := 0; i < b.N; i++ {
+		t.MustPut(keys[i], hashes[i])
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		t.Del(keys[i])
+	}
+}
+
+func BenchmarkBinHash(b *testing.B) {
+	t := NewBin()
+	keys := make([][]byte, 0, b.N)
+	hashes := make([][]byte, 0, b.N)
+	for i := 0; i < b.N; i++ {
+		key := make([]byte, 32)
+		hash := make([]byte, 32)
+		_, _ = rand.Read(key)
+		_, _ = rand.Read(hash)
+		keys = append(keys, key)
+		hashes = append(hashes, hash)
+	}
+	for i := 0; i < b.N; i++ {
+		t.MustPut(keys[i], hashes[i])
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		t.Hash()
+	}
+}
